Move cross_fields tests to their own file.
